pyproject.toml
[project]
name = "comic_pile"
version = "0.1.0"
description = "Dice-Driven Comic Reading Tracker - FastAPI + HTMX + Tailwind CSS"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.12.0",
    "jinja2>=3.1.0",
    "pydantic>=2.0.0",
    "python-multipart>=0.0.6",
    "psycopg[binary]>=3.1.0",
    "slowapi>=0.1.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "pytest-xdist>=3.0.0",
    "ruff>=0.1.0",
    "pyright>=1.1.0",
    "faker>=20.0.0",
    "httpx>=0.24.0",
    "playwright==1.48.0",
    "pytest-playwright==0.7.0",
    "uvicorn>=0.30.0",
    "PyGithub>=2.0.0",
]

[dependency-groups]
dev = [
    "playwright==1.48.0",
    "pytest-asyncio>=1.3.0",
    "pytest-playwright==0.7.0",
]

[tool.node-dependencies]
dev = [
    "eslint>=9.0.0",
    "htmlhint>=1.0.0",
]

[tool.pytest.ini_options]
testpaths = ["tests", "tests_e2e"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--cov=comic_pile",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-fail-under=90",
]
asyncio_mode = "auto"
markers = [
    "integration: marks tests as integration tests (requires Playwright)",
]

[tool.coverage.run]
source = ["app", "comic_pile"]
omit = [
    "*/tests/*",
    "*/test_*.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [".venv", "__pycache__", "*.pyc", ".git", "node_modules", "alembic/versions/*.py"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "C4", "D", "ANN401"]
ignore = ["D203", "D213", "E501", "B008", "I001"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pyright]
include = ["app", "comic_pile", "tests"]
exclude = ["**/node_modules", "**/__pycache__", ".venv", "alembic/versions/*.py"]
venvPath = "."
venv = ".venv"
typeStubPaths = []

[tool.pytest.markers]
integration = "marks tests as integration tests using Playwright"
